requirements.txt
dnspython==1.16.0
PySocks==1.6.8
stem==1.7.1
